install if not exists pip

+++ run.bat
@@ -1,5 +1,5 @@
 @echo on
-choco install python --version 3.7.2
+if not exist "C:/Python37/Scripts/pip.exe" choco install python --version 3.7.2
 set PATH=c:/python37/Scripts;c:/python37/;%PATH%
 set PYTHON=C:/Python37/python.exe
 set PIP=C:/Python37/Scripts/pip.exe
